+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added plugin version to the system report log
+
 ### Fixed
 - Fixed a bug where manual invoices for backend created orders did not go to processing state.
 - Fixed a bug where filling in certain wrong values for the Date of birth field for AfterPay and in3 caused an error