+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+- Added automatic cancellation MultiSafepay pretratransaction for non-paid canceled orders.
 - Added plugin version to the system report log
 
 ### Fixed